requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.1.6
django-cors-headers==3.14.0
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
et-xmlfile==1.1.0
filelock==3.10.7
idna==3.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
openpyxl==3.1.2
packaging==23.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
psycopg2==2.9.6
PyJWT==2.6.0
pyngrok==6.0.0
pypng==0.20220715.0
python-decouple==3.7
python-version==0.0.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rest-framework-simplejwt==0.0.2
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.0
yarg==0.1.9
